==> 01_zri_multifamily_cleaning.ipynb <==
import pandas as pd
import numpy as np

import pandas as pd
import numpy as np

def time_series_gaps(df, ts_start_col):
    '''
    Returns a dictionary containing:
        keys: the row indices of any row with missing data
        values: numeric horizontal (column) start and end indices of any periods of missing data
    
    time series data must be running row-wise for the function to work
    
    ts_start_col is the numeric index of the column in the dataframe at which the time series begins
    the function assumes the time series runs from ts_start_col all the way to the right side of the dataframe
    '''
    gaps = {}
    for i in range(df.shape[0]):
        curr_gaps = []
        curr_ts = df.iloc[i,:]
        if curr_ts.isnull().any():
            j = ts_start_col
            while j < len(curr_ts):
                if pd.isnull(curr_ts[j]):
                    gap_start = int(j)
                    gap_end = None
                    k = int(j)
                    while gap_end is None:
                        if k == len(curr_ts) - 1:
                            if pd.isnull(curr_ts[k]):
                                gap_end = int(k)
                                curr_gaps.append((gap_start, gap_end))
                                j = int(k)+1
                            else:   
                                gap_end = int(k) - 1
                                curr_gaps.append((gap_start, gap_end))
                                j = int(k)+1
                        elif pd.isnull(curr_ts[k]):
                            k+=1
                        else:
                            gap_end = int(k) - 1
                            curr_gaps.append((gap_start, gap_end))
                            j = int(k)
                else:
                    j += 1
        if curr_gaps != []:
            gaps[i] = curr_gaps
    return gaps

def get_rejected_gaps(gaps, gap_limit):
    '''
    takes gaps dictionary generated by time_series_gaps, returns list of row indices that contain gaps >= gap_limit
    '''
    over_limit = []
    for key, val in gaps.items():
        if len(val) > 1:
            for i in range(len(val)):
                width = val[i][1] - val[i][0] + 1
                if width >= gap_limit:
                    over_limit.append(key)
        else:
            width = val[0][1] - val[0][0] + 1
            if width >= gap_limit:
                over_limit.append(key)
    return list(set(over_limit))

multi = pd.read_csv('./../data/Zip_Zri_MultiFamilyResidenceRental.csv')

type(multi['RegionName'][0])

multi

multi['RegionName'] = multi['RegionName'].map(lambda x: str(x))

multi['RegionName'].map(lambda x: len(x)).value_counts()

multi['RegionName'] = multi['RegionName'].map(lambda x: '0'+x if len(x)<5 else x)

multi.rename(columns = {'RegionName':'zip'}, inplace = True)

multi = multi.iloc[:,1:] # drop RegionID column, just a unique identifier

multi

multi = pd.concat([multi.iloc[:,0:6], multi.iloc[:,46:]], axis = 1)

multi

gaps = time_series_gaps(multi, 7)

rej_gaps = get_rejected_gaps(gaps, 6)

multi2 = multi.drop(rej_gaps, axis = 0)

multi2

multi2.columns[7:]

multi2 = pd.melt(multi2, id_vars = multi2.columns[0:6], value_vars = multi2.columns[6:], var_name = 'year-month', value_name = 'zri')

missing_zri = multi2[multi2['zri'].isnull()]['zip'].unique()

missing_zri

list(multi2.loc[multi2['zip']=='78212','zri'])

multi2.loc[multi2['zip']=='78212','zri'].interpolate(method = 'ffill').interpolate(method = 'bfill')

for item in missing_zri:
    temp = multi2.loc[multi2['zip']==item,'zri']
    multi2.loc[multi2['zip']==item,'zri'] = temp.interpolate(method = 'linear').interpolate(method = 'ffill').interpolate(method = 'bfill')

multi2['zri'].isnull().value_counts()

multi2 = multi2[multi2['year-month']!='2020-01']

multi2.drop(columns = ['SizeRank'], inplace = True)

multi2['year-month'].value_counts()

multi2 = multi2.sort_values(['zip', 'year-month'])

multi2.shape

multi2.to_csv('./../data/zri_multifamily_v2.csv', index = False)
