==> 02_ACS_5yr_zip_merging.ipynb <==
import pandas as pd
import numpy as np

_2014 = pd.read_csv('./data/acs_2014_zip_5yr.csv')

_2015 = pd.read_csv('./data/acs_2015_zip_5yr.csv')

_2016 = pd.read_csv('./data/acs_2016_zip_5yr.csv')

_2017 = pd.read_csv('./data/acs_2017_zip_5yr.csv')

_2018 = pd.read_csv('./data/acs_2018_zip_5yr.csv')

_2013 = pd.read_csv('./data/acs_2013_zip_5yr.csv')

_2012 = pd.read_csv('./data/acs_2012_zip_5yr.csv')

_2011 = pd.read_csv('./data/acs_2011_zip_5yr.csv')

years = [_2011, _2012, _2013, _2014, _2015, _2016, _2017, _2018]

_2018['do_date'] = [20142018]*33120

for item in years:
    print(item['do_date'][0])
    print(type(item['do_date'][0]))

for item in years:
    if 'geoid' in item.columns:
        print(item['do_date'][0], 'geoid')
        print(item['geoid'][0])
        print(type(item['geoid'][0]))
    if 'geo_id' in item.columns:
        print(item['do_date'][0], 'geo_id')
        print(item['geo_id'][0])
        print(type(item['geo_id'][0]))

for i in range(len(years)):
    print(len(years[i].columns))

for col in _2011.columns:
    if col not in _2015.columns:
        print(col, 1)
    if col not in _2018.columns:
        print(col, 2)

merged = pd.concat(years, axis = 0)

merged.drop(columns = 'geoid', axis = 1, inplace = True)

merged = pd.concat([merged.loc[:,merged.columns[0]], merged.loc[:,merged.columns[-1]], merged.loc[:, merged.columns[1:-1]]], axis = 1)

merged = merged.sort_values(['geo_id', 'do_date'])

merged.to_csv('acs_5yr_merged.csv', index = False)
