==> 04_ACS_feat_eng.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('./../data/acs_5yr_merged_v3.csv')

df

df['zip'] = df['zip'].map(lambda x: str(x))

df['zip'] = df['zip'].map(lambda x: '0'+x if len(x) < 5 else x)

df['zip'].map(lambda x: len(x)).value_counts()

df

df.rename(columns = {'year':'year_usable'}, inplace = True)

df['year_usable'][0]

df['year_usable'] = df['year_usable'] + 2

# ——
# average commute time
# education: 
# education per income?
# -
# median year structure built
# average units per building
# —
# race: white black asian hispanic indian,… populations
# ———
# vacant housing units for rent (as a percentage?)
# age: pre-working age, millenial, working age, senior
# education: high-school, bachelor, advanced (as all perentages of pop)
# gini_index

list(df.columns)

df2 = pd.DataFrame()
df2['zip'] = df['zip']
df2['census_period'] = df['do_date']
df2['year_usable'] = df['year_usable']

pre_working_cols = [
    'male_under_5',
    'male_5_to_9',
    'male_10_to_14',
    'male_15_to_17',
    'female_under_5',
    'female_5_to_9',
    'female_10_to_14',
    'female_15_to_17'
]

adult_cols = [
    'male_18_to_19',
    'male_20',
    'male_21',
    'male_22_to_24',
    'male_25_to_29',
    'male_30_to_34',
    'male_35_to_39',
    'female_18_to_19',
    'female_20',
    'female_21',
    'female_22_to_24',
    'female_25_to_29',
    'female_30_to_34',
    'female_35_to_39'
]

older_adult_cols = [
    'female_40_to_44',
    'female_45_to_49',
    'female_50_to_54',
    'female_55_to_59',
    'female_60_to_61',
    'female_62_to_64',
    'male_40_to_44',
    'male_45_to_49',
    'male_50_to_54',
    'male_55_to_59',
    'male_60_61',
    'male_62_64'
]

senior_cols = [
    'female_65_to_66',
    'female_67_to_69',
    'female_70_to_74',
    'female_75_to_79',
    'female_80_to_84',
    'female_85_and_over',
    'male_65_to_66',
    'male_67_to_69',
    'male_70_to_74',
    'male_75_to_79',
    'male_80_to_84',
    'male_85_and_over'
]

df2['percent_white'] = df['white_pop']/df['total_pop']
df2['percent_black'] = df['black_pop']/df['total_pop']
df2['percent_asian'] = df['asian_pop']/df['total_pop']
df2['percent_hispanic'] = df['hispanic_pop']/df['total_pop']
df2['percent_native_am'] = df['amerindian_pop']/df['total_pop']
df2['percent_other_race'] = df['other_race_pop']/df['total_pop']

df2['percent_0_17'] = df[pre_working_cols].sum(axis=1)/df['total_pop']
df2['percent_18_39'] = df[adult_cols].sum(axis=1)/df['total_pop']
df2['percent_40_64'] = df[older_adult_cols].sum(axis=1)/df['total_pop']
df2['percent_65+'] = df[senior_cols].sum(axis=1)/df['total_pop']

# trying to measure demand of rental units
df2['percent_rental_units_vacant'] = df['vacant_housing_units_for_rent']/(df['vacant_housing_units_for_rent']+df['housing_units_renter_occupied'])
df2['percent_rental_units_occupied'] = df['housing_units_renter_occupied']/(df['vacant_housing_units_for_rent']+df['housing_units_renter_occupied'])

total_dwellings = df[[ 'dwellings_1_units_detached',
                                         'dwellings_1_units_attached',
                                         'dwellings_2_units',
                                         'dwellings_3_to_4_units',
                                         'dwellings_5_to_9_units',
                                         'dwellings_10_to_19_units',
                                         'dwellings_20_to_49_units',
                                         'dwellings_50_or_more_units']].sum(axis = 'columns')

df['derived_pop_over_25']= df[['male_25_to_29',
 'male_30_to_34',
 'male_35_to_39',
 'male_40_to_44',
 'male_45_to_49',
 'male_50_to_54',
 'male_55_to_59',
 'male_60_61',
 'male_62_64',
 'male_65_to_66',
 'male_67_to_69',
 'male_70_to_74',
 'male_75_to_79',
 'male_80_to_84',
 'male_85_and_over',
 'female_25_to_29',
 'female_30_to_34',
 'female_35_to_39',
 'female_40_to_44',
 'female_45_to_49',
 'female_50_to_54',
 'female_55_to_59',
 'female_60_to_61',
 'female_62_to_64',
 'female_65_to_66',
 'female_67_to_69',
 'female_70_to_74',
 'female_75_to_79',
 'female_80_to_84',
 'female_85_and_over']].sum(axis='columns')

#EDUCATION
df2['percent_graduate_deg'] = df['graduate_professional_degree']/df['total_pop'] #population 
df2['percent_bachelors'] = df['bachelors_degree_2']/df['total_pop'] #population 
df2['percent_associates'] = df['some_college_and_associates_degree']/df['derived_pop_over_25'] #25 and over 
df2['percent_highschool'] = df['high_school_including_ged']/df['derived_pop_over_25']  #25 and over 
df2['percent_less_highschool'] = df['less_than_high_school_graduate']/df['derived_pop_over_25']  #25 and over

#COMMUTE
df2['percent_commute_public_transport'] = df['commuters_by_public_transportation']/df['commuters_16_over']

df2['percent_commute_less_30'] = (df[['commute_5_9_mins',
                                   'commute_10_14_mins',
                                   'commute_15_19_mins',
                                   'commute_20_24_mins',
                                   'commute_25_29_mins']].sum(axis='columns'))/df['commuters_16_over']

df2['percent_buildings_less_10_units'] = (df[['dwellings_1_units_detached',
                                         'dwellings_1_units_attached',
                                         'dwellings_2_units',
                                         'dwellings_3_to_4_units',
                                         'dwellings_5_to_9_units']].sum(axis='columns')) / total_dwellings

                                        
df2['percent_buildings_10_19_units'] = df['dwellings_10_to_19_units'] / total_dwellings
                                        

df2['percent_buildings_20_49_units'] = df['dwellings_20_to_49_units'] / total_dwellings
                                        

df2['percent_buildings_50+_units'] = df['dwellings_50_or_more_units'] / total_dwellings

df2['percent_commute_30_to_59'] = (df[['commute_30_34_mins',
                                     'commute_35_44_mins',
                                     'commute_35_39_mins',
                                     'commute_40_44_mins',
                                     'commute_45_59_mins']].sum(axis='columns'))/df['commuters_16_over']

df2['percent_commute_60_to_89'] = df['commute_60_89_mins']/df['commuters_16_over']

df2['percent_commute_90_more'] = df['commute_90_more_mins']/df['commuters_16_over']

# MOVERS
# moved to this city from another city 
df2['percent_new_city'] = df['different_house_year_ago_different_city']/df['total_pop']

# moved to a new unit in this same city
df2['percent_new_unit'] = df['different_house_year_ago_same_city']/df['total_pop']

# OWNER OCCUPIED
df2['percent_units_owner_occupied'] = df['owner_occupied_housing_units']/df['housing_units']

# features to keep/keep in slightly modified form

# change median year building built to median building age
df2['median_building_age'] = (df['year_usable']-2) - df['median_year_structure_built']
df2['income_per_capita'] = df['income_per_capita']
df2['poverty_rate'] = df['poverty']/df['pop_determined_poverty_status']
df2['total_pop'] = df['total_pop']
df2['percent_workforce_unemployed'] = df['unemployed_pop']/df['civilian_labor_force']
df2['percent_work_from_home'] = df['worked_at_home']/df['civilian_labor_force']
df2['median_age'] = df['median_age']
df2['percent_female'] = df['female_pop']/df['total_pop']
df2['gini_index'] = df['gini_index']
df2['percent_not_us_citizen'] = df['not_us_citizen_pop']/df['total_pop']

df2

df2.to_csv('./../data/acs_engineered_features.csv', index=False)
