==> 05_merging_acs_bikeshare_zri.ipynb <==
import pandas as pd
import numpy as np
import plotly.express as px
import plotly.graph_objects as go
from datetime import datetime
import pmdarima as pm
from pmdarima.utils import diff
from sklearn.metrics import mean_squared_error
from pmdarima.metrics import smape
import pickle
import sklearn

pd.set_option('display.max_rows', 500)

# Importing and preparing zri and acs data for merging

zri = pd.read_csv('./../data/zri_multifamily_v2.csv')

zri['zip'] = zri['zip'].map(lambda x: str(x))

zri['zip'].map(lambda x: len(x)).value_counts()

zri['zip'] = zri['zip'].map(lambda x: '0' + x if len(x)<5 else x)

zri['zip'].map(lambda x: len(x)).value_counts()

zri['year']=zri['year-month'].map(lambda x: int(x[0:4]))

zri['month']=zri['year-month'].map(lambda x: int(x[5:]))

years = range(2014,2020)

months = range(1, 13)

datetime_col = [datetime(year=i, month=j, day=1) for i in years for j in months]

datetime_col = datetime_col*1302

zri['datetime'] = datetime_col

zri.drop(columns = 'year-month', inplace = True)

zri

acs = pd.read_csv('./../data/acs_engineered_features.csv')

acs['zip'] = acs['zip'].map(lambda x: str(x))

acs['zip'].map(lambda x: len(x)).value_counts()

acs['zip'] = acs['zip'].map(lambda x: '0' + x if len(x)<5 else x)

acs['zip'].map(lambda x: len(x)).value_counts()

acs.drop(columns = 'census_period', axis = 1, inplace = True)

acs

# Imputing ACS monthly data

acs_zips = acs['zip'].unique()

years = acs['year_usable'].unique()

zri_zips = zri['zip'].unique()

zri_zips

acs_zips_in_zri = []

for item in acs_zips:
    if item in zri_zips:
        acs_zips_in_zri.append(item)

acs = acs.set_index('zip')

acs = acs.loc[acs_zips_in_zri,:]

acs = acs.reset_index(drop=False)

acs

acs = acs.loc[acs.index.repeat(12)].reset_index(drop=True)

acs['month'] = 0

months = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12] * len(acs_zips_in_zri) * len(years)

acs['month'] = months

acs.loc[acs['month']!=12, acs.columns[2:-1]] = np.nan

acs

for item in acs_zips_in_zri:
    temp = acs.loc[acs['zip']==item,:]
    for col in temp.columns[2:-1]:
        temp[col] = temp[col].interpolate(method='linear')
    acs.loc[acs['zip']==item,:] = temp

acs

merged = pd.merge(zri, acs, left_on = ['zip','year', 'month'], right_on = ['zip','year_usable', 'month'], how = 'left')

null_check = {}
for col in merged.columns:
    if merged[col].isnull().any():
        null_check[col] = merged[col].isnull().sum()
null_check

merged= merged[merged['zip']!= '11249']

null_check = {}
for col in merged.columns:
    if merged[col].isnull().any():
        null_check[col] = merged[col].isnull().sum()
null_check

merged['zip'].nunique()

df = merged.drop(columns = ['year_usable'])

df

bikeshare = pd.read_csv('./../data/bikeshare_engineered_features.csv', index_col = 0)

bikeshare['month'] = bikeshare['year-month'].map(lambda x: int(x[5:7]))

bikeshare.drop(columns = ['City', 'State', 'Metro', 'CountyName', 'year-month', 'zri'], inplace = True)

bikeshare['zip'] = bikeshare['zip'].map(lambda x: str(x))

bikeshare['zip'] = bikeshare['zip'].map(lambda x: '0' +x if len(x)<5 else x)

df = pd.merge(df, bikeshare, on = ('zip', 'year', 'month'), how = 'left')

df

df['has_bike_sharing'] = df['has_bike_sharing'].map(lambda x: 1 if x==True else 0)

# df.to_csv('./../data/zri_acs_bikeshare_merged.csv', index = 0)
