==> 06_google_trends_data_collection.ipynb <==
from datetime import datetime, timedelta
import pandas as pd
import numpy as np
from pytrends.request import TrendReq
pytrend = TrendReq()

kw_list=["Abortion",'Gun control', 'Starbucks', 'Universal basic income', 'Affirmative action']

metro_dict = {
    'Albany-Schenectady-Troy':'Albany-Schenectady-Troy NY',
    'Albuquerque':'Albuquerque-Santa Fe NM',
    'Allentown-Bethlehem-Easton': 'Wilkes Barre-Scranton PA',
    'Anchorage':'Anchorage AK',
    'Athens-Clarke County': 'Macon GA',
    'Atlanta-Sandy Springs-Roswell':'Atlanta GA',
    'Augusta-Waterville':'Augusta GA',
    'Austin-Round Rock':'Austin TX',
    'Bakersfield':'Bakersfield CA',
    'Baltimore-Columbia-Towson':'Baltimore MD',
    'Bangor':'Bangor ME',
    'Batavia':'Rochester NY',
    'Binghamton':'Binghamton NY',
    'Bloomington':  'Indianapolis IN',
    'Boston-Cambridge-Newton' : 'Boston MA-Manchester NH',
    'Boulder':'Denver CO',
    'Bozeman': 'Butte-Bozeman MT',
    'Bridgeport-Stamford-Norwalk' : 'Hartford & New Haven CT',
    'Brownsville-Harlingen': 'Corpus Christi TX',
    'Buffalo-Cheektowaga-Niagara Falls' : 'Buffalo NY',
    'Canton-Massillon':'Cleveland-Akron (Canton) OH',
    'Cape Coral-Fort Myers': 'Ft. Myers-Naples FL',
    'Charleston-North Charleston':'Charleston SC',
    'Charlotte-Concord-Gastonia':'Charlotte NC',
    'Charlottesville':'Charlottesville VA',
    'Chicago-Naperville-Elgin' : 'Chicago IL',
    'Chico' : 'Chico-Redding CA',
    'Cincinnati' : 'Cincinnati OH',
    'Clarksville' : 'Nashville TN',
    'Cleveland-Elyria' : 'Cleveland-Akron (Canton) OH',
    'College Station-Bryan' : 'Waco-Temple-Bryan TX',
    'Colorado Springs' : 'Colorado Springs-Pueblo CO',
    'ColumbiaSC' : 'Columbia SC',
    'ColumbiaMO' : 'Columbia-Jefferson City MO',
    'Columbus' : 'Columbus OH',
    'Concord' :  'Boston MA-Manchester NH',
    'Dallas-Fort Worth-Arlington' : 'Dallas-Ft. Worth TX',
    'Dayton' : 'Dayton OH',
    'Denver-Aurora-Lakewood' : 'Denver CO',
    'Detroit-Warren-Dearborn' : 'Detroit MI',
    'Dubuque' : 'Cedar Rapids-Waterloo-Iowa City & Dubuque IA',
    'Durham-Chapel Hill' : 'Raleigh-Durham (Fayetteville) NC',
    'El Paso' : 'El Paso TX',
    'Erie' : 'Erie PA',
    'Eugene' : 'Eugene OR',
    'Fargo' : 'Fargo-Valley City ND',
    'Fayetteville-Springdale-Rogers' : 'Ft. Smith-Fayetteville-Springdale-Rogers AR',
    'Findlay' : 'Lima OH',
    'Greenville-Anderson-Mauldin' : 'Greenville-Spartanburg SC-Asheville NC-Anderson SC',
    'Harrisburg-Carlisle' : 'Harrisburg-Lancaster-Lebanon-York PA',
    'Hartford-West Hartford-East Hartford' : 'Hartford & New Haven CT',
    'Helena' : 'Helena MT',
    'Houston-The Woodlands-Sugar Land' : 'Houston TX',
    'Indianapolis-Carmel-Anderson' : 'Indianapolis IN',
    'Ithaca' : 'Elmira NY',
    'Jacksonville' : 'Jacksonville FL',
    'Jamestown-Dunkirk-Fredonia' : 'Erie PA',
    'Joplin' : 'Joplin MO-Pittsburg KS',
    'Kalispell' : 'Missoula MT',
    'Kansas City' : 'Kansas City MO',
    'Killeen-Temple' : 'Waco-Temple-Bryan TX',
    'Kingston' : 'Albany-Schenectady-Troy NY',
    'La Crosse-Onalaska' : 'La Crosse-Eau Claire WI',
    'Lancaster' : 'Harrisburg-Lancaster-Lebanon-York PA',
    'Las Vegas-Henderson-Paradise' : 'Las Vegas NV',
    'Lewiston-Auburn' : 'Portland-Auburn ME',
    'Lexington-Fayette' : 'Lexington KY',
    'Little Rock-North Little Rock-Conway' : 'Little Rock-Pine Bluff AR',
    'Los Angeles-Long Beach-Anaheim' : 'Los Angeles CA',
    'Madison' : 'Madison WI',
    'Manchester-Nashua' : 'Boston MA-Manchester NH',
    'Manhattan' : 'Topeka KS', # Manhattan, Kansas
    'Manitowoc' : 'Green Bay-Appleton WI',
    'Medford' : 'Medford-Klamath Falls OR',
    'Memphis' : 'Memphis TN',
    'Miami-Fort Lauderdale-West Palm Beach' : 'Miami-Ft. Lauderdale FL',
    'Milwaukee-Waukesha-West Allis' : 'Milwaukee WI',
    'Minneapolis-St. Paul-Bloomington' : 'Minneapolis-St. Paul MN',
    'Missoula' : 'Missoula MT',
    'Morgantown' :  'Clarksburg-Weston WV', #morgantown, WV
    'Nashville-Davidson--Murfreesboro--Franklin' : 'Nashville TN',
    'New Haven-Milford' : 'Hartford & New Haven CT',
    'New York-Newark-Jersey City' : 'New York NY',
    'North Port-Sarasota-Bradenton' : 'Tampa-St. Petersburg (Sarasota) FL',
    'Norwich-New London' : 'Hartford & New Haven CT', # norwich, CT
    'Ogden-Clearfield' : 'Salt Lake City UT', #utah
    'Oklahoma City' : 'Oklahoma City OK',
    'Orlando-Kissimmee-Sanford' : 'Orlando-Daytona Beach-Melbourne FL',
    'Oshkosh-Neenah' : 'Green Bay-Appleton WI', # oshkosh wi
    'Oxnard-Thousand Oaks-Ventura' : 'Los Angeles CA', #ca
    'Palm Bay-Melbourne-Titusville' : 'Orlando-Daytona Beach-Melbourne FL',
    'Philadelphia-Camden-Wilmington' : 'Philadelphia PA',
    'Phoenix-Mesa-Scottsdale' : 'Phoenix AZ',
    'Pittsburgh' : 'Pittsburgh PA',
    'Portland-South Portland' : 'Portland-Auburn ME',
    'Portland-Vancouver-Hillsboro' : 'Portland OR',
    'Prescott' : 'Phoenix AZ', #presscot AZ
    'Providence-Warwick' : 'Providence RI-New Bedford MA',
    'Racine' : 'Milwaukee WI', # racine wi
    'Raleigh' : 'Raleigh-Durham (Fayetteville) NC',
    'Reno' : 'Reno NV',
    'Richmond' : 'Richmond-Petersburg VA',
    'Riverside-San Bernardino-Ontario' : 'Los Angeles CA', # ca
    'Rochester' : 'Rochester NY',
    'Rockford' : 'Rockford IL',
    'Sacramento--Roseville--Arden-Arcade' : 'Sacramento-Stockton-Modesto CA',
    'Salinas' : 'Monterey-Salinas CA', # salinas CA
    'Salt Lake City' : 'Salt Lake City UT',
    'San Antonio-New Braunfels' : 'San Antonio TX',
    'San Diego-Carlsbad' : 'San Diego CA',
    'San Francisco-Oakland-Hayward' : 'San Francisco-Oakland-San Jose CA',
    'San Jose-Sunnyvale-Santa Clara' : 'San Francisco-Oakland-San Jose CA',
    'Savannah' : 'Savannah GA',
    'Seattle-Tacoma-Bellevue' : 'Seattle-Tacoma WA',
    'Spokane-Spokane Valley' : 'Spokane WA',
    'Springfield' : 'Springfield-Holyoke MA',
    'St. Louis' : 'St. Louis MO',
    'Stockton-Lodi' : 'Sacramento-Stockton-Modesto CA',
    'Syracuse' : 'Syracuse NY',
    'Tampa-St. Petersburg-Clearwater' : 'Tampa-St. Petersburg (Sarasota) FL',
    'Topeka' : 'Topeka KS',
    'Torrington' : 'Hartford & New Haven CT', # torrington ct
    'Tucson' : 'Tucson (Sierra Vista) AZ',
    'Tyler' : 'Tyler-Longview(Lufkin & Nacogdoches) TX',
    'Urban Honolulu' : 'Honolulu HI',
    'Utica-Rome' : 'Utica NY',
    'Virginia Beach-Norfolk-Newport News' : 'Norfolk-Portsmouth-Newport News VA',
    'Warren' : 'Erie PA', # warren pa
    'Washington-Arlington-Alexandria' : 'Washington DC (Hagerstown MD)',
    'Watertown-Fort Drum' : 'Watertown NY',
    'Wheeling' : 'Wheeling WV-Steubenville OH',
    'Wichita' : 'Wichita-Hutchinson KS',
    'Williamsport' : 'Wilkes Barre-Scranton PA', # williamsport pa
    'Worcester' : 'Boston MA-Manchester NH',
    'York-Hanover' : 'Harrisburg-Lancaster-Lebanon-York PA'
}

trendslist=pd.read_csv('../data/googletrends_countries.csv')
trendslist=trendslist[trendslist['country_code']=='US']
trendslist.reset_index(drop=True, inplace = True)
metro_codes = trendslist.iloc[20266:,:].reset_index(drop=True)

metros_to_match = []

for item in metro_dict.values():
    metros_to_match.append(item.lower())

keep_rows = []

for i in range(metro_codes.shape[0]):
    if metro_codes.loc[i, 'name'].lower() in metros_to_match:
        keep_rows.append(i)

metro_codes = metro_codes.loc[keep_rows, :]

metro_code_dict = dict(zip(metro_codes['sub_code'], metro_codes['name'].map(lambda x: x.lower())))

metro_list = metro_codes['sub_code'].map(lambda x: 'US-'+x).tolist()

df2=pd.DataFrame(columns=['date','isPartial','metro_code'])
df2=df2.set_index('date')

for item in metro_list:
    pytrend.build_payload(kw_list, cat=None, timeframe='2014-01-01 2020-01-01', geo=item) 

    df = pytrend.interest_over_time()
    df['metro_code']=item
    df2=df2.append(df)

df3 = df2.copy()

df2['metro_code'] = df2['metro_code'].map(lambda x: x[3:])

df2['Metro'] = df2['metro_code'].map(lambda x: metro_code_dict[x])

df2 = df2.reset_index(drop=False).sort_values(['Metro', 'date'])

zillow = pd.read_csv('./../data/zri_acs_bikeshare_merged.csv')

zillow = pd.DataFrame(zillow['Metro']).drop_duplicates().reset_index(drop=True).rename(columns = {'Metro':'Metro_zillow'})

zillow['Metro'] = zillow['Metro_zillow'].map(lambda x: metro_dict[x].lower())

merged = pd.merge(zillow, df2, on = 'Metro', how = 'left')

merged = merged.drop(columns = ['Metro', 'metro_code']).rename(columns = {'Metro_zillow': 'Metro', 'date':'datetime'})

for item in kw_list:
    merged[f'rel_pop_{item}'] = np.nan

for metro in merged['Metro'].unique():
    temp = merged.loc[merged['Metro']==metro,:]
    max_col = None
    max_pop = 0
    for col in temp.columns[3:3+len(kw_list)]:
        if max(temp[col]) > max_pop:
            max_col = col
            max_pop = max(temp[col])

    for col in temp.columns[3:3+len(kw_list)]:
        temp[f'rel_pop_{col}'] = max(temp[col])
        temp[col] = temp[col] * 100 / max(temp[col])
    merged.loc[merged['Metro']==metro,:] = temp

merged

merged.to_csv('./google_terms_political.csv', index = 0)
