==> 07_merging_gtrends_dbs_tax.ipynb <==
import pandas as pd
import numpy as np
from sklearn.preprocessing import StandardScaler

guns = pd.read_csv('./../data/google_trend/google_terms_guns.csv')

jobs = pd.read_csv('./../data/google_trend/google_terms_jobs.csv')

lgbt = pd.read_csv('./../data/google_trend/google_terms_lgbt.csv')

social = pd.read_csv('./../data/google_trend/google_terms_social.csv')

travels = pd.read_csv('./../data/google_trend/google_terms_travels.csv')

woty = pd.read_csv('./../data/google_trend/google_terms_woty.csv')

disaster = pd.read_csv('./../data/google_trend/google_terms_disasters.csv')

gentrification = pd.read_csv('./../data/google_trend/google_terms_gentrification2.csv')

multifamily = pd.read_csv('./../data/google_trend/google_terms_multifamily_for_rent.csv')

homes = pd.read_csv('./../data/google_trend/google_terms_homes_for_rent.csv')

apartment = pd.read_csv('./../data/google_trend/google_terms_apartment_for_rent.csv')

df_list = [guns, jobs, lgbt, social, travels, woty, disaster, gentrification, multifamily, homes, apartment]
i=0
ss = StandardScaler()
for item in df_list:
    item.columns = item.columns.map(lambda x: x.lower())
    item.drop(columns = item.columns[8:], inplace=True)
    for col in item.columns[3:]:
        temp = pd.DataFrame(item[col])
        temp = ss.fit_transform(temp)
        item[col] = temp
    item.drop(columns = 'ispartial', inplace = True)
    if i != 0:
        item.drop(columns = ['metro', 'datetime'], inplace = True)
    i+=1

full_trends = pd.concat(df_list, axis = 1).fillna(0)

full_trends_backup = full_trends.copy()

full_trends['starbucks2'] = full_trends['starbucks'].mean(axis=1)

full_trends.drop(columns = 'starbucks', inplace = True)

full_trends.rename(columns = {'starbucks2' : 'starbucks'}, inplace = True)

full_trends.drop(columns = ['tender age shelter', 'tourists attraction', 'gun licence'], inplace = True)

full_trends.rename(columns = {'metro':'Metro'}, inplace = True)

full_trends['year_usable'] = full_trends['datetime'].map(lambda x: int(x[0:4])+1)

full_trends['month'] = full_trends['datetime'].map(lambda x: int(x[5:7]))

# full_trends.to_csv('./../data/google_trends_full.csv', index = False)

df = pd.read_csv('./../data/zri_acs_bikeshare_merged.csv', index_col = 0)

df.reset_index(drop=False, inplace = True)

df = pd.merge(df, full_trends, how = 'left', left_on = ['Metro', 'year', 'month'], right_on = ['Metro', 'year_usable', 'month'])

df

df.drop(columns = 'datetime_y', inplace = True)

df.rename(columns = {'datetime_x':'datetime'}, inplace = True)

bds = pd.read_csv('./../data/bds_updated.csv', index_col = 0)

merged = pd.merge(df, bds, how = 'left', left_on = ['zip', 'State', 'CountyName', 'zri', 'datetime'], right_on = ['zip', 'State', 'CountyName', 'zri', 'year-month'])

merged

merged.drop(columns = ['City_y', 'Metro_y', 'year_y', 'year-month'], inplace = True)

merged.rename(columns = {'City_x': 'City', 'Metro_x': 'Metro', 'year_x':'year'}, inplace = True)

merged

taxes = pd.read_csv('./../data/final_TAXES_by_zip.csv')

zips = list(merged['zip'].unique())

taxes = taxes.rename(columns = {'ZIPCODE': 'zip'})

taxes = taxes[taxes['zip'].isin(zips)]

taxes['year_usable'] = taxes['year']+1

taxes.drop(columns = 'year', inplace = True)

taxes

merged2 = pd.merge(merged, taxes, left_on = ['zip', 'year', 'month'], right_on = ['zip', 'year_usable', 'month'], how = 'left')

merged2

merged2.drop(columns = ['year_usable_x', 'year_usable_y'], inplace = True)

merged2

merged2.to_csv('./../data/full_dataset_unscaled.csv', index = False)
